condensates: extract legacy RGIE support primitives
