+++ src/exogibbs/optimize/legacy_condensate/__init__.py
@@ -1,0 +1,1 @@
+"""Focused implementation modules for explicit legacy condensate routes."""
