+++ src/exogibbs/optimize/pipm_rgie_cond.py
@@ -9,7 +9,6 @@ from jax import debug as jdebug
 from jax.lax import while_loop
 from jax.lax import cond
 from jax.lax import fori_loop
-from jax.scipy.linalg import cho_factor, cho_solve
 from jax.scipy.linalg import lu_factor, lu_solve
 from typing import Dict, Tuple, Optional, Sequence, Any
 from time import perf_counter
@@ -17,6 +16,20 @@ from time import perf_counter
 from exogibbs.api.chemistry import ThermoState
 from exogibbs.optimize.core import _A_diagn_At
 from exogibbs.optimize.core import _compute_gk
+from exogibbs.optimize.legacy_condensate.rgie_helpers import (
+    build_rgie_condensate_init_from_policy,
+    summarize_rgie_inactive_driving,
+)
+from exogibbs.optimize.legacy_condensate.rgie_reduced_system import (
+    DEFAULT_REDUCED_SOLVER,
+    DEFAULT_REGULARIZATION_MODE,
+    DEFAULT_REGULARIZATION_STRENGTH,
+    _assemble_reduced_system_terms,
+    _recompute_pi_for_residual,
+    _regularize_q_block,
+    _solve_reduced_gibbs_iteration_equations_cond_with_metrics,
+    solve_reduced_gibbs_iteration_equations_cond,
+)
 from exogibbs.optimize.minimize import solve_gibbs_iteration_equations
 from exogibbs.optimize.pipm_gie_cond import (
     solve_gibbs_iteration_equations_cond as solve_full_gibbs_iteration_equations_cond,
@@ -28,9 +41,6 @@ from exogibbs.optimize.stepsize import stepsize_cond_heurstic
 from exogibbs.optimize.stepsize import stepsize_sk
 from exogibbs.optimize.stepsize import LOG_S_MAX
 
-DEFAULT_REDUCED_SOLVER = "augmented_lu_row_scaled"
-DEFAULT_REGULARIZATION_MODE = "none"
-DEFAULT_REGULARIZATION_STRENGTH = 0.0
 KL_DENSITY_GAUGE_P0_CGS = 1.0e6
 KL_DENSITY_GAUGE_K_B_CGS = 1.380649e-16
 
@@ -1276,159 +1286,6 @@ def compute_rgie_lambda_cap_policy(
     }
 
 
-def build_rgie_condensate_init_from_policy(
-    epsilon: float,
-    support_indices: jnp.ndarray,
-    startup_policy: str,
-    *,
-    driving: Optional[jnp.ndarray] = None,
-    m0: Optional[float] = None,
-    r0: Optional[float] = None,
-    top_k: Optional[int] = None,
-    tiny_fallback: float = 1.0e-30,
-    dtype: Optional[jnp.dtype] = None,
-) -> jnp.ndarray:
-    """Build a diagnostic-only RGIE condensate initialization from a startup policy.
-
-    The returned vector is defined only on the currently supported condensates.
-    This helper does not change production defaults; callers must explicitly opt
-    into using it.
-    """
-
-    support_indices = jnp.asarray(support_indices)
-    if support_indices.ndim != 1:
-        raise ValueError("support_indices must be a one-dimensional index array.")
-
-    if dtype is None:
-        dtype = jnp.float64
-
-    eps = jnp.asarray(epsilon, dtype=dtype)
-    n_support = int(support_indices.shape[0])
-    if n_support == 0:
-        return jnp.zeros((0,), dtype=dtype)
-
-    if tiny_fallback <= 0.0:
-        raise ValueError("tiny_fallback must be positive.")
-    fallback_ln_m0 = jnp.log(jnp.asarray(tiny_fallback, dtype=dtype))
-
-    if startup_policy == "absolute_uniform_m0":
-        if m0 is None or m0 <= 0.0:
-            raise ValueError("absolute_uniform_m0 requires a positive m0.")
-        return jnp.full((n_support,), jnp.log(jnp.asarray(m0, dtype=dtype)), dtype=dtype)
-
-    target_ln_m0 = None
-    if startup_policy in (
-        "ratio_uniform_r0",
-        "ratio_positive_driving_r0",
-        "ratio_topk_positive_driving_r0",
-    ):
-        if r0 is None or r0 <= 0.0:
-            raise ValueError(f"{startup_policy} requires a positive r0.")
-        target_ln_m0 = eps + jnp.log(jnp.asarray(r0, dtype=dtype))
-
-    if startup_policy == "ratio_uniform_r0":
-        return jnp.full((n_support,), target_ln_m0, dtype=dtype)
-
-    if startup_policy in ("ratio_positive_driving_r0", "ratio_topk_positive_driving_r0"):
-        if driving is None:
-            raise ValueError(f"{startup_policy} requires driving values.")
-        driving = jnp.asarray(driving, dtype=dtype)
-        if driving.shape != (n_support,):
-            raise ValueError(
-                "driving must have the same shape as the supported condensate block "
-                f"(got {driving.shape}, expected {(n_support,)})."
-            )
-        positive = driving > 0.0
-
-        if startup_policy == "ratio_positive_driving_r0":
-            selected = positive
-        else:
-            if top_k is None:
-                raise ValueError("ratio_topk_positive_driving_r0 requires top_k.")
-            if top_k < 0:
-                raise ValueError("top_k must be non-negative.")
-            if top_k == 0:
-                selected = jnp.zeros((n_support,), dtype=bool)
-            else:
-                safe_driving = jnp.where(positive, driving, -jnp.inf)
-                ranked = jnp.argsort(-safe_driving)
-                top_indices = ranked[: min(top_k, n_support)]
-                selected = jnp.zeros((n_support,), dtype=bool).at[top_indices].set(True)
-                selected = selected & positive
-
-        return jnp.where(selected, target_ln_m0, fallback_ln_m0)
-
-    raise ValueError(
-        "Unknown startup_policy "
-        f"'{startup_policy}'. Expected one of "
-        "('absolute_uniform_m0', 'ratio_uniform_r0', "
-        "'ratio_positive_driving_r0', 'ratio_topk_positive_driving_r0')."
-    )
-
-
-def summarize_rgie_inactive_driving(
-    full_driving: jnp.ndarray,
-    support_indices: jnp.ndarray,
-    *,
-    condensate_species_names: Optional[Sequence[str]] = None,
-    top_k: int = 5,
-) -> Dict[str, Any]:
-    """Summarize inactive-driving violations for a current support."""
-
-    driving = np.asarray(jax.device_get(full_driving), dtype=np.float64)
-    support = np.asarray(jax.device_get(support_indices), dtype=np.int64)
-    n_cond = int(driving.shape[0])
-    support_mask = np.zeros((n_cond,), dtype=bool)
-    if support.size:
-        support_mask[support] = True
-    inactive_indices = np.nonzero(~support_mask)[0]
-
-    if inactive_indices.size == 0:
-        return {
-            "max_positive_inactive_driving": 0.0,
-            "inactive_positive_count": 0,
-            "top_inactive_indices": [],
-            "top_inactive_names": [],
-            "top_inactive_driving": [],
-            "top_positive_inactive_indices": [],
-        }
-
-    inactive_driving = driving[inactive_indices]
-    positive_mask = inactive_driving > 0.0
-    positive_indices = inactive_indices[positive_mask]
-    positive_driving = inactive_driving[positive_mask]
-
-    if positive_indices.size == 0:
-        return {
-            "max_positive_inactive_driving": 0.0,
-            "inactive_positive_count": 0,
-            "top_inactive_indices": [],
-            "top_inactive_names": [],
-            "top_inactive_driving": [],
-            "top_positive_inactive_indices": [],
-        }
-
-    ranked_order = np.argsort(-positive_driving)
-    ranked_positive = positive_indices[ranked_order]
-    top_indices = ranked_positive[: min(int(top_k), int(ranked_positive.shape[0]))]
-    top_driving = driving[top_indices]
-    if condensate_species_names is None:
-        top_names = [str(int(index)) for index in top_indices]
-    else:
-        top_names = [
-            str(condensate_species_names[int(index)]) for index in top_indices
-        ]
-
-    return {
-        "max_positive_inactive_driving": float(np.max(positive_driving)),
-        "inactive_positive_count": int(positive_indices.shape[0]),
-        "top_inactive_indices": [int(index) for index in top_indices],
-        "top_inactive_names": top_names,
-        "top_inactive_driving": [float(value) for value in top_driving],
-        "top_positive_inactive_indices": [int(index) for index in ranked_positive],
-    }
-
-
 def build_rgie_support_candidate_indices(
     support_indices: jnp.ndarray,
     *,
@@ -1499,63 +1356,6 @@ def build_rgie_support_candidate_indices(
         "inactive_positive_ranked": ranked_add,
         "weak_active_ranked": weak_active_ranked,
     }
-
-
-def _assemble_reduced_system_terms(
-    nk: jnp.ndarray,
-    mk: jnp.ndarray,
-    ntotk: float,
-    formula_matrix: jnp.ndarray,
-    formula_matrix_cond: jnp.ndarray,
-    b: jnp.ndarray,
-    gk: jnp.ndarray,
-    bk: jnp.ndarray,
-    hvector_cond: jnp.ndarray,
-    sk: jnp.ndarray,
-) -> Dict[str, jnp.ndarray]:
-    """Assemble the reduced condensed system once for backend experiments."""
-
-    resn = jnp.sum(nk) - ntotk
-    Qk = _A_diagn_At(nk, formula_matrix) + _A_diagn_At(sk, formula_matrix_cond)
-    Angk = formula_matrix @ (gk * nk)
-    ngk = jnp.dot(nk, gk)
-    delta_bk_hat = b - (bk + formula_matrix_cond @ mk)
-    condvec = formula_matrix_cond @ (sk * hvector_cond - mk)
-    rhs = Angk + condvec + delta_bk_hat
-    scalar_rhs = ngk - resn
-    assemble_mat = jnp.block([[Qk, bk[:, None]], [bk[None, :], jnp.array([[resn]])]])
-    assemble_vec = jnp.concatenate([rhs, jnp.array([scalar_rhs])])
-    return {
-        "resn": resn,
-        "Qk": Qk,
-        "rhs": rhs,
-        "scalar_rhs": scalar_rhs,
-        "assemble_mat": assemble_mat,
-        "assemble_vec": assemble_vec,
-    }
-
-
-def _regularize_q_block(
-    q_block: jnp.ndarray,
-    regularization_mode: str,
-    regularization_strength: float,
-) -> Tuple[jnp.ndarray, jnp.ndarray]:
-    """Apply explicit optional regularization to the Q block only."""
-
-    dtype = q_block.dtype
-    reg_used = (
-        jnp.asarray(regularization_strength, dtype=dtype)
-        if regularization_mode == "diag_shift"
-        else jnp.asarray(0.0, dtype=dtype)
-    )
-    if regularization_mode == "none":
-        return q_block, reg_used
-    if regularization_mode != "diag_shift":
-        raise ValueError(
-            f"Unknown regularization_mode '{regularization_mode}'. Expected 'none' or 'diag_shift'."
-        )
-    eye = jnp.eye(q_block.shape[0], dtype=dtype)
-    return q_block + reg_used * eye, reg_used
 
 
 def _diagnostic_json_array(value: Any) -> list[Any]:
@@ -12329,217 +12129,6 @@ def _normalize_exact_input_bundle_context(
     }
 
 
-def solve_reduced_gibbs_iteration_equations_cond(
-    nk: jnp.ndarray,
-    mk: jnp.ndarray,
-    ntotk: float,
-    formula_matrix: jnp.ndarray,
-    formula_matrix_cond: jnp.ndarray,
-    b: jnp.ndarray,
-    gk: jnp.ndarray,
-    bk: jnp.ndarray,
-    hvector_cond: jnp.ndarray,
-    sk: jnp.ndarray,
-    *,
-    reduced_solver: str = DEFAULT_REDUCED_SOLVER,
-    regularization_mode: str = DEFAULT_REGULARIZATION_MODE,
-    regularization_strength: float = DEFAULT_REGULARIZATION_STRENGTH,
-) -> Tuple[jnp.ndarray, float]:
-    """
-        Solve the reduced Gibbs iteration equations with condensates using the Lagrange multipliers.
-        This function computes the matrix and vector to solve the system of equations
-        that arises from the Gibbs energy minimization problem.
-
-        Args:
-            nk: number of species vector (n_species,) for k-th iteration.
-            mk: number of condensed species vector (n_cond,) for k-th iteration.
-            ntotk: Total number of species for k-th iteration.
-            formula_matrix: Gas Formula matrix for stoichiometric constraints (n_elements, n_species).
-            formula_matrix_cond: Condensates Formula matrix for stoichiometric constraints (n_elements, n_cond).
-            b: Element abundance vector (n_elements, ).
-            gk: gk vector (n_species,) for k-th iteration.
-            bk: (gas) formula_matrix @ nk vector (n_elements, ).
-            hvector_cond: chemical_potentials for condensates divided by RT (n_cond, )
-            sk: mk^2/nu (n_cond, )
-
-        Returns:
-            Tuple containing:
-                - The pi vector (nelements, ).fastchem_elements = list(gas.elements)
-                    element_indices = jnp.array([fastchem_elements.index(e) for e in elements])
-
-                - The update of the  log total number of species (delta_ln_ntot).
-    """
-
-    pi_vector, delta_ln_ntot, _metrics = _solve_reduced_gibbs_iteration_equations_cond_with_metrics(
-        nk,
-        mk,
-        ntotk,
-        formula_matrix,
-        formula_matrix_cond,
-        b,
-        gk,
-        bk,
-        hvector_cond,
-        sk,
-        reduced_solver=reduced_solver,
-        regularization_mode=regularization_mode,
-        regularization_strength=regularization_strength,
-    )
-    return pi_vector, delta_ln_ntot
-
-
-def _solve_reduced_gibbs_iteration_equations_cond_with_metrics(
-    nk: jnp.ndarray,
-    mk: jnp.ndarray,
-    ntotk: float,
-    formula_matrix: jnp.ndarray,
-    formula_matrix_cond: jnp.ndarray,
-    b: jnp.ndarray,
-    gk: jnp.ndarray,
-    bk: jnp.ndarray,
-    hvector_cond: jnp.ndarray,
-    sk: jnp.ndarray,
-    *,
-    reduced_solver: str = DEFAULT_REDUCED_SOLVER,
-    regularization_mode: str = DEFAULT_REGULARIZATION_MODE,
-    regularization_strength: float = DEFAULT_REGULARIZATION_STRENGTH,
-    include_system_trace: bool = False,
-) -> Tuple[jnp.ndarray, float, Dict[str, jnp.ndarray]]:
-    """Solve the reduced system and expose scale metrics for debugging."""
-
-    assembled = _assemble_reduced_system_terms(
-        nk,
-        mk,
-        ntotk,
-        formula_matrix,
-        formula_matrix_cond,
-        b,
-        gk,
-        bk,
-        hvector_cond,
-        sk,
-    )
-    resn = assembled["resn"]
-    q_block = assembled["Qk"]
-    rhs = assembled["rhs"]
-    scalar_rhs = assembled["scalar_rhs"]
-    assemble_mat = assembled["assemble_mat"]
-    assemble_vec = assembled["assemble_vec"]
-
-    reg_q_block, reg_used = _regularize_q_block(
-        q_block,
-        regularization_mode,
-        regularization_strength,
-    )
-
-    row_scale = jnp.maximum(jnp.max(jnp.abs(assemble_mat), axis=1, keepdims=True), 1.0)
-    col_scale = jnp.maximum(jnp.max(jnp.abs(assemble_mat), axis=0, keepdims=True), 1.0)
-    row_scale_flat = row_scale[:, 0]
-    col_scale_flat = col_scale[0, :]
-
-    if reduced_solver == "augmented_lu_row_scaled":
-        solve_mat = assemble_mat / row_scale
-        solve_vec = assemble_vec / row_scale[:, 0]
-        lu, piv = lu_factor(solve_mat)
-        assemble_variable = lu_solve((lu, piv), solve_vec)
-        factorization_succeeded = jnp.all(jnp.isfinite(assemble_variable))
-    elif reduced_solver == "augmented_lu_rowcol_scaled":
-        reg_assemble_mat = jnp.block(
-            [[reg_q_block, bk[:, None]], [bk[None, :], jnp.array([[resn]])]]
-        )
-        row_scale = jnp.maximum(jnp.max(jnp.abs(reg_assemble_mat), axis=1, keepdims=True), 1.0)
-        solve_mat_row = reg_assemble_mat / row_scale
-        col_scale = jnp.maximum(jnp.max(jnp.abs(solve_mat_row), axis=0, keepdims=True), 1.0)
-        solve_mat = solve_mat_row / col_scale
-        solve_vec = assemble_vec / row_scale[:, 0]
-        lu, piv = lu_factor(solve_mat)
-        scaled_solution = lu_solve((lu, piv), solve_vec)
-        assemble_variable = scaled_solution / col_scale[0, :]
-        factorization_succeeded = jnp.all(jnp.isfinite(assemble_variable))
-        row_scale_flat = row_scale[:, 0]
-        col_scale_flat = col_scale[0, :]
-        assemble_mat = reg_assemble_mat
-    elif reduced_solver == "schur_cholesky_reg":
-        cho, lower = cho_factor(reg_q_block)
-        rhs_pair = jnp.stack((rhs, bk), axis=1)
-        solved_pair = cho_solve((cho, lower), rhs_pair)
-        qinv_rhs = solved_pair[:, 0]
-        qinv_bk = solved_pair[:, 1]
-        schur = resn - jnp.vdot(bk, qinv_bk)
-        schur_safe = jnp.where(
-            jnp.abs(schur) < 1.0e-300,
-            jnp.where(schur < 0.0, -1.0e-300, 1.0e-300),
-            schur,
-        )
-        delta_ln_ntot = (scalar_rhs - jnp.vdot(bk, qinv_rhs)) / schur_safe
-        pi_vector = qinv_rhs - qinv_bk * delta_ln_ntot
-        assemble_variable = jnp.concatenate([pi_vector, jnp.array([delta_ln_ntot])])
-        factorization_succeeded = jnp.all(jnp.isfinite(assemble_variable))
-    else:
-        raise ValueError(
-            "Unknown reduced_solver "
-            f"'{reduced_solver}'. Expected one of "
-            "('augmented_lu_row_scaled', 'augmented_lu_rowcol_scaled', 'schur_cholesky_reg')."
-        )
-
-    metrics = {
-        "reduced_solver_backend": reduced_solver,
-        "reduced_factorization_succeeded": factorization_succeeded,
-        "reduced_regularization_mode": regularization_mode,
-        "reduced_regularization_strength": jnp.asarray(regularization_strength, dtype=q_block.dtype),
-        "reduced_regularization_used": reg_used,
-        "reduced_resn": resn,
-        "reduced_row_scale_min": jnp.min(row_scale_flat),
-        "reduced_row_scale_max": jnp.max(row_scale_flat),
-        "reduced_row_scale_ratio": jnp.max(row_scale_flat)
-        / jnp.maximum(jnp.min(row_scale_flat), 1.0e-300),
-        "reduced_col_scale_min": jnp.min(col_scale_flat),
-        "reduced_col_scale_max": jnp.max(col_scale_flat),
-        "reduced_col_scale_ratio": jnp.max(col_scale_flat)
-        / jnp.maximum(jnp.min(col_scale_flat), 1.0e-300),
-        "reduced_mat_maxabs": jnp.max(jnp.abs(assemble_mat)),
-        "reduced_vec_maxabs": jnp.max(jnp.abs(assemble_vec)),
-        "reduced_qk_maxabs": jnp.max(jnp.abs(q_block)),
-        "reduced_qk_diag_min": jnp.min(jnp.diag(q_block)),
-        "reduced_qk_diag_max": jnp.max(jnp.diag(q_block)),
-    }
-    if include_system_trace:
-        metrics.update(
-            {
-                "diagnostic_only_reduced_system_trace": jnp.asarray(True),
-                "reduced_jacobian_matrix_before_scaling": assemble_mat,
-                "reduced_rhs_vector_before_scaling": assemble_vec,
-                "reduced_row_scaling_vector": row_scale_flat,
-                "reduced_col_scaling_vector": col_scale_flat,
-                "reduced_solve_matrix": solve_mat,
-                "reduced_solve_rhs_vector": solve_vec,
-                "reduced_raw_solver_result_vector": assemble_variable,
-                "reduced_q_block": q_block,
-                "reduced_rhs_element_block": rhs,
-                "reduced_scalar_rhs": jnp.asarray(scalar_rhs),
-                "reduced_element_condensate_block_before_scaling": (
-                    jnp.asarray(formula_matrix_cond, dtype=jnp.float64)
-                    * jnp.asarray(mk, dtype=jnp.float64)[None, :]
-                ),
-                "reduced_element_condensate_block_scaled_by_reduced_row_scaling": (
-                    (
-                        jnp.asarray(formula_matrix_cond, dtype=jnp.float64)
-                        * jnp.asarray(mk, dtype=jnp.float64)[None, :]
-                    )
-                    / jnp.maximum(row_scale_flat[:-1], 1.0e-300)[:, None]
-                ),
-                "reduced_element_condensate_owner_mk_vector": jnp.asarray(
-                    mk,
-                    dtype=jnp.float64,
-                ),
-                "reduced_system_trace_basis": jnp.asarray(1, dtype=jnp.int32),
-            }
-        )
-    return assemble_variable[:-1], assemble_variable[-1], metrics
-
-
-
-
 def _compute_residuals(
     nk: jnp.ndarray,
     mk: jnp.ndarray,
@@ -14824,48 +14413,6 @@ def diagnose_pdipm_vs_pipm_fixed_epsilon_trajectories(
         },
         "pdipm_runs": pdipm_runs,
     }
-
-
-def _recompute_pi_for_residual(
-    nk: jnp.ndarray,
-    mk: jnp.ndarray,
-    ntot: float,
-    formula_matrix: jnp.ndarray,
-    formula_matrix_cond: jnp.ndarray,
-    b: jnp.ndarray,
-    gk: jnp.ndarray,
-    hvector_cond: jnp.ndarray,
-    epsilon: float,
-    *,
-    reduced_solver: str = DEFAULT_REDUCED_SOLVER,
-    regularization_mode: str = DEFAULT_REGULARIZATION_MODE,
-    regularization_strength: float = DEFAULT_REGULARIZATION_STRENGTH,
-) -> jnp.ndarray:
-    """Re-solve the reduced system on the current state for residual evaluation only.
-
-    This solve is intentionally separate from the earlier solve that produced the
-    update direction. It is not fed back into the primal update; it is only used
-    to evaluate a post-update barrier residual on a self-consistent state.
-    """
-
-    bk = formula_matrix @ nk
-    sk = mk * mk * jnp.exp(-epsilon)
-    pi_vector, _delta_ln_ntot = solve_reduced_gibbs_iteration_equations_cond(
-        nk,
-        mk,
-        ntot,
-        formula_matrix,
-        formula_matrix_cond,
-        b,
-        gk,
-        bk,
-        hvector_cond,
-        sk,
-        reduced_solver=reduced_solver,
-        regularization_mode=regularization_mode,
-        regularization_strength=regularization_strength,
-    )
-    return pi_vector
 
 
 def _compute_iteration_step_metrics(
